+++ R/mighty_domain.R
@@ -1,7 +1,8 @@
 #' Mighty Domain
 #'
 #' @description
-#' `mighty_domain()` provides a robust way of working with ADaM metadata in the `{mighty}` framework.
+#' `mighty_domain()` provides a robust way of working with ADaM metadata
+#' in the `{mighty}` framework.
 #'
 #' A new object is initialized by supplying either an existing yaml metadata
 #' file or an in-memory `list` of the same content.
@@ -13,8 +14,8 @@
 #' * `help("rows")`
 #'
 #' `mighty_domain()` inherits from `S7schema::S7schema()` and the yaml file is
-#' automatically validated when loaded. The helper functions above also always validates
-#' the new configuration before returning.
+#' automatically validated when loaded. The helper functions above also always
+#' validates the new configuration before returning.
 #'
 #' You can at anytime validate an object by calling `validate()` and use
 #' `write_config()` to save it as a yaml file again.